backend/requirements.txt
flask==3.1.0
flask-cors==6.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
pyjwt==2.9.0
requests==2.32.3
schedule==1.2.2
reportlab==4.4.4
fpdf2==2.7.9
pandas==2.3.3
